refactor/request data v3 (#130)
This PR continues #128 and ensures `RequestPipeline` can be shared over
many requests unless a local configuration is provided.

Similar to `RequestData` being shared in #128.

This includes further refactorings:

- `DateTimeProvider` is only provided externally once (on `NodePool`)
and that instance is used everywhere. Before it could be set seperately
on `NodePool` and `TransportConfiguration`, not by design but by
necessity.

- `RequestPipeline` is no longer disposable (we didn't actually disposed
anything).

- A new type exists `Auditor` this is now explicitly passed to the
methods that need it on `RequestPipeline`. It implements
`IReadOnlyCollection<Audit>` and is exposed as such to users.

- This PR also merges `DefaultRequestPipeline` and `RequestPipeline`
into one.